requirements/common.txt
alembic==1.8.0
blinker==1.4
click==8.1.3
dominate==2.6.0
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
python-dotenv==0.20.0
SQLAlchemy==1.4.39
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
